Base: Changing integers of PID to floats to fix controller behavior

+++ orion_base/orion_ctl_micro_ros/lib/pid/pid.cpp
@@ -18,35 +18,35 @@ namespace diff
         if(!this->enabled_)
         {
             // If so, take advantage to reset the PID.
-            if(this->last_input_ != 0)
+            if(this->last_input_ != 0.0f)
             {
                 reset(enc_count);
             }
             return;
         }
 
-        // Set up input
-        int input = enc_count - this->last_enc_count_;
+        // Set up input (encoder ticks since last call)
+        float input = (float)(enc_count - this->last_enc_count_);
 
         // Determinate the error
-        long err = this->setpoint_ - input;
+        float err = this->setpoint_ - input;
 
         // Calculate output
         // Considering PID with a Ko value to avoid too much increment
-        long output = (this->kp_ * err - this->kd_ * (input - this->last_input_)
+        float output = (this->kp_ * err - this->kd_ * (input - this->last_input_)
             + this->integral_term_) / this->ko_;
 
         // Sum previous output
         output += this->last_output_;
 
         // Clamp to avoid saturation
-        if(output > this->pwm_max_)
+        if(output > (float)this->pwm_max_)
         {
-            output = this->pwm_max_;
+            output = (float)this->pwm_max_;
         }
-        else if (output < this->pwm_min_)
+        else if (output < (float)this->pwm_min_)
         {
-            output = this->pwm_min_;
+            output = (float)this->pwm_min_;
         }
         else
         {
@@ -55,7 +55,7 @@ namespace diff
         }
 
         // Update value of the commanded output
-        computed_output = output;
+        computed_output = (int)output;
 
         // Prepare values for the next iteration
         this->last_enc_count_ = enc_count;
@@ -100,11 +100,11 @@ namespace diff
      */
     void ControlPID::reset(int enc_count)
     {
-        this->setpoint_ = 0;
-        this->integral_term_ = 0;
+        this->setpoint_ = 0.0f;
+        this->integral_term_ = 0.0f;
         this->last_enc_count_ = enc_count;
-        this->last_input_ = 0;
-        this->last_output_ = 0;
+        this->last_input_ = 0.0f;
+        this->last_output_ = 0.0f;
 
     } // void ControlPID::reset()
 
@@ -113,7 +113,7 @@ namespace diff
      *
      * @param setpoint Encoder count to achieve
      */
-    void ControlPID::setSetpoint(int setpoint)
+    void ControlPID::setSetpoint(float setpoint)
     {
         this->setpoint_ = setpoint;
 
@@ -127,7 +127,7 @@ namespace diff
      * @param ki Integral constant
      * @param ko Saturation constant
      */
-    void ControlPID::setTunings(int kp, int kd, int ki, int ko)
+    void ControlPID::setTunings(float kp, float kd, float ki, float ko)
     {
         this->kp_ = kp;
         this->kd_ = kd;
